add back support for directly specifying a nameserver
In v3.0 the code got reworked to query zone data from proper DNS
resolvers instead of always querying the manually configured
(authoritative) nameserver. This is generally good and fixed a lot of
use cases with delegated zones, but it broke two other use cases in
turn:
    - using a hidden primary, where the only nameserver capable of
    updating the zone is not publicly listed as an authority
    - using different IPs internally and externaly, because the public
      IPs may change, the internal IPs are used for additional security,
      or both

To re-enable those use cases in the least invasive way, I added the
ability to optionally specify one single "HiddenPrimary" IP which
prevents the NS queries from happening and will be used in place of
their result. The resolver is still used for all the non-updating
queries.